prak 6 & 7 done

+++ prak6/percobaan3.c
@@ -2,18 +2,18 @@
 
 void main()
 {
-    int angka_pertama, angka_kedua, menu, hasil;
+    float angka_pertama, angka_kedua, menu, hasil;
     printf("Masukkan bilangan pertama: ");
-    scanf("%d", &angka_pertama);
+    scanf("%f", &angka_pertama);
     printf("Masukkan bilangan kedua: ");
-    scanf("%d", &angka_kedua);
+    scanf("%f", &angka_kedua);
     printf("Menu Matematika\n");
     printf("1. Penjumlahan\n");
     printf("2. Pengurangan\n");
     printf("3. Pembagian\n");
     printf("4. Perkalian\n");
     printf("Masukkan pilihan anda: ");
-    scanf("%d", &menu);
+    scanf("%f", &menu);
 
     if (menu == 1)
     {
@@ -36,5 +36,5 @@ void main()
         printf("Menu tidak valid!\n");
         return;
     }
-    printf("Hasil dari operasi tersebut = %d\n", hasil);
+    printf("Hasil dari operasi tersebut = %.1f\n", hasil);
 }
